refactored methods
