+++ packages/trader/src/Modules/CFD/Components/cfd-demo-account-display.jsx
@@ -1,6 +1,5 @@
 import React from 'react';
 import { localize } from '@deriv/translations';
-import { CFD_PLATFORMS, isLandingCompanyEnabled } from '@deriv/shared';
 import specifications from 'Modules/CFD/Constants/cfd-specifications';
 import { CFDAccountCard } from './cfd-account-card.jsx';
 import { general_messages } from '../Constants/cfd-shared-strings';
@@ -15,7 +14,9 @@ const CFDDemoAccountDisplay = ({
     standpoint,
     is_loading,
     is_logged_in,
-    landing_companies,
+    isSyntheticCardVisible,
+    isFinancialCardVisible,
+    isFinancialStpCardVisible,
     onSelectAccount,
     openAccountTransfer,
     platform,
@@ -33,27 +34,6 @@ const CFDDemoAccountDisplay = ({
                 type: 'financial',
             });
         }
-    };
-
-    const isSyntheticCardVisible = () => {
-        return isLandingCompanyEnabled({ landing_companies, platform, type: 'gaming' }) || !is_logged_in;
-    };
-
-    const isFinancialCardVisible = () => {
-        return (
-            !is_logged_in ||
-            isLandingCompanyEnabled({
-                landing_companies,
-                platform,
-                type: 'financial',
-            })
-        );
-    };
-
-    const isFinancialStpCardVisible = () => {
-        return (
-            (landing_companies?.mt_financial_company?.financial_stp || !is_logged_in) && platform === CFD_PLATFORMS.MT5
-        );
     };
 
     const financial_specs = React.useMemo(() => {
