+++ packages/trader/src/Modules/CFD/Containers/cfd-dashboard.jsx
@@ -2,7 +2,14 @@ import React from 'react';
 import { withRouter } from 'react-router';
 import { Redirect } from 'react-router-dom';
 import { DesktopWrapper, Icon, MobileWrapper, Tabs, PageError, Loading, Text } from '@deriv/components';
-import { isEmptyObject, isMobile, routes, getCFDPlatformLabel, CFD_PLATFORMS } from '@deriv/shared';
+import {
+    isEmptyObject,
+    isMobile,
+    routes,
+    getCFDPlatformLabel,
+    CFD_PLATFORMS,
+    isLandingCompanyEnabled,
+} from '@deriv/shared';
 import { Localize, localize } from '@deriv/translations';
 import { ResetTradingPasswordModal } from '@deriv/account';
 import { connect } from 'Stores/connect';
@@ -186,6 +193,28 @@ class CFDDashboard extends React.Component {
             dxtrade_verification_code,
         } = this.props;
 
+        const isSyntheticCardVisible = () => {
+            return isLandingCompanyEnabled({ landing_companies, platform, type: 'gaming' }) || !is_logged_in;
+        };
+
+        const isFinancialCardVisible = () => {
+            return (
+                !is_logged_in ||
+                isLandingCompanyEnabled({
+                    landing_companies,
+                    platform,
+                    type: 'financial',
+                })
+            );
+        };
+
+        const isFinancialStpCardVisible = () => {
+            return (
+                (landing_companies?.mt_financial_company?.financial_stp || !is_logged_in) &&
+                platform === CFD_PLATFORMS.MT5
+            );
+        };
+
         const should_show_missing_real_account =
             !is_eu && is_logged_in && !has_real_account && upgradeable_landing_companies?.length > 0;
         if ((!country && is_logged_in) || is_logging_in) return <Loading />; // Wait for country name to be loaded before rendering
@@ -291,6 +320,9 @@ class CFDDashboard extends React.Component {
                                                     is_pending_authentication={is_pending_authentication}
                                                     is_fully_authenticated={is_fully_authenticated}
                                                     is_virtual={is_virtual}
+                                                    isSyntheticCardVisible={isSyntheticCardVisible}
+                                                    isFinancialCardVisible={isFinancialCardVisible}
+                                                    isFinancialStpCardVisible={isFinancialStpCardVisible}
                                                     openAccountTransfer={this.openAccountTransfer}
                                                     openPasswordManager={this.togglePasswordManagerModal}
                                                     openPasswordModal={this.openRealPasswordModal}
@@ -321,6 +353,9 @@ class CFDDashboard extends React.Component {
                                             openAccountNeededModal={openAccountNeededModal}
                                             standpoint={standpoint}
                                             is_loading={is_loading}
+                                            isSyntheticCardVisible={isSyntheticCardVisible}
+                                            isFinancialCardVisible={isFinancialCardVisible}
+                                            isFinancialStpCardVisible={isFinancialStpCardVisible}
                                             has_cfd_account={has_cfd_account}
                                             current_list={current_list}
                                             onSelectAccount={createCFDAccount}
